+++ joern-cli/frontends/jimple2cpg/src/main/scala/io/joern/jimple2cpg/passes/SootAstCreationPass.scala
@@ -1,6 +1,7 @@
 package io.joern.jimple2cpg.passes
 
 import io.joern.jimple2cpg.Config
+import io.joern.jimple2cpg.astcreation.AstCreator
 import io.joern.x2cpg.datastructures.Global
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.passes.ConcurrentWriterCpgPass
